requirements.txt
groq>=0.11.0
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
scikit-learn>=1.4.0